refactor(core): centralise provider adapter spawn helpers
Each of the three provider adapters (claude, codex, cursor) shipped its own
copy of the same two blocks:
- a `--version` detection probe, ~30 lines of identical Promise wiring;
- a ~130-line async-iterator that spawns the CLI, parses each stdout line,
  and yields TurnEvents with leak-free cleanup.

Centralise both into providers/shared/spawn-stream.ts:
- detectBinary(binary, spawnFn) returns DetectResult
- spawnLineStream(binary, args, spawnFn, { parseLine, parseCtx, onClose? })
  returns AsyncIterable<TurnEvent>

Each adapter now builds its CLI args, calls the helper, and is left with
only its provider-specific concerns (claude permission flags, codex synthetic
done event, cursor --force flag).

Net: -340 lines, single source of truth for child-process bookkeeping.
