+++ packages/core/src/providers/claude/adapter.ts
@@ -1,5 +1,4 @@
-import { spawn, type ChildProcess } from 'node:child_process';
-import { createInterface } from 'node:readline';
+import { spawn } from 'node:child_process';
 import type {
   DetectResult,
   IsoDateTime,
@@ -9,6 +8,7 @@ import type {
   TurnEvent,
   TurnRequest,
 } from '@kay-am/types';
+import { detectBinary, spawnLineStream } from '../shared/spawn-stream';
 import { computeCostUsd } from './cost';
 import { parseStreamJsonLine } from './parser';
 
@@ -45,37 +45,7 @@ export class ClaudeAdapter implements ProviderAdapter {
   }
 
   async detect(): Promise<DetectResult> {
-    return new Promise((resolve) => {
-      const child = this.spawnFn(this.binary, ['--version'], {
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      let stdout = '';
-      child.stdout?.on('data', (chunk: Buffer) => {
-        stdout += chunk.toString('utf8');
-      });
-      child.on('error', (err) => {
-        resolve({
-          kind: 'missing',
-          binary: this.binary,
-          reason: err.message,
-        });
-      });
-      child.on('close', (code) => {
-        if (code === 0) {
-          resolve({
-            kind: 'available',
-            binary: this.binary,
-            version: stdout.trim(),
-          });
-        } else {
-          resolve({
-            kind: 'missing',
-            binary: this.binary,
-            reason: `exited with code ${code}`,
-          });
-        }
-      });
-    });
+    return detectBinary(this.binary, this.spawnFn);
   }
 
   cost(usage: ProviderUsage, model: string): number {
@@ -83,123 +53,36 @@ export class ClaudeAdapter implements ProviderAdapter {
   }
 
   spawn(request: TurnRequest): AsyncIterable<TurnEvent> {
-    return spawnClaude(this.binary, this.spawnFn, this.now, this.onUnknown, request);
-  }
-}
+    const prompt = `${request.systemPrompt}\n\n${request.userMessage}`.trim();
+    const flags = request.permissionFlags;
+    const mode = flags?.mode ?? 'default';
 
-async function* spawnClaude(
-  binary: string,
-  spawnFn: typeof spawn,
-  now: () => IsoDateTime,
-  onUnknown: (type: string, payload: unknown) => void,
-  request: TurnRequest,
-): AsyncIterable<TurnEvent> {
-  const prompt = `${request.systemPrompt}\n\n${request.userMessage}`.trim();
-  const flags = request.permissionFlags;
-  const mode = flags?.mode ?? 'default';
+    const args = [
+      '-p',
+      prompt,
+      '--output-format',
+      'stream-json',
+      '--working-dir',
+      request.workingDir,
+      '--model',
+      request.model,
+      '--permission-mode',
+      mode,
+    ];
 
-  const args = [
-    '-p',
-    prompt,
-    '--output-format',
-    'stream-json',
-    '--working-dir',
-    request.workingDir,
-    '--model',
-    request.model,
-    '--permission-mode',
-    mode,
-  ];
+    const allowedTools = flags?.allowedTools ?? [];
+    const disallowedTools = flags?.disallowedTools ?? [];
 
-  const allowedTools = flags?.allowedTools ?? [];
-  const disallowedTools = flags?.disallowedTools ?? [];
-
-  if (allowedTools.length > 0) {
-    args.push('--allowedTools', allowedTools.join(','));
-  }
-  if (disallowedTools.length > 0) {
-    args.push('--disallowedTools', disallowedTools.join(','));
-  }
-
-  const child: ChildProcess = spawnFn(binary, args, {
-    stdio: ['ignore', 'pipe', 'pipe'],
-  });
-
-  if (!child.stdout) {
-    throw new Error('claude CLI started without stdout');
-  }
-
-  const queue: TurnEvent[] = [];
-  let resolver: ((value: IteratorResult<TurnEvent>) => void) | null = null;
-  let rejector: ((err: unknown) => void) | null = null;
-  let ended = false;
-  let error: unknown = null;
-
-  const ctx = { runId: request.runId, now, onUnknown };
-
-  const flush = () => {
-    if (resolver && queue.length > 0) {
-      const value = queue.shift()!;
-      const r = resolver;
-      resolver = null;
-      r({ value, done: false });
-    } else if (resolver && ended) {
-      const r = resolver;
-      resolver = null;
-      if (error) {
-        const rej = rejector;
-        rejector = null;
-        rej?.(error);
-      } else {
-        r({ value: undefined, done: true });
-      }
+    if (allowedTools.length > 0) {
+      args.push('--allowedTools', allowedTools.join(','));
     }
-  };
-
-  const lineReader = createInterface({ input: child.stdout });
-
-  lineReader.on('line', (line) => {
-    const events = parseStreamJsonLine(line, ctx);
-    for (const event of events) queue.push(event);
-    flush();
-  });
-
-  lineReader.on('close', () => {
-    ended = true;
-    flush();
-  });
-
-  child.on('error', (err) => {
-    error = err;
-    ended = true;
-    flush();
-  });
-
-  child.stderr?.on('data', () => {
-    // captured but not surfaced as TurnEvent for v0.1
-  });
-
-  try {
-    while (true) {
-      if (queue.length > 0) {
-        yield queue.shift()!;
-        continue;
-      }
-      if (ended) {
-        if (error) throw error;
-        return;
-      }
-      const value = await new Promise<IteratorResult<TurnEvent>>((resolve, reject) => {
-        resolver = resolve;
-        rejector = reject;
-      });
-      if (value.done) return;
-      yield value.value;
+    if (disallowedTools.length > 0) {
+      args.push('--disallowedTools', disallowedTools.join(','));
     }
-  } finally {
-    if (!child.killed && child.exitCode === null) {
-      child.kill('SIGTERM');
-    }
-    lineReader.close();
+
+    return spawnLineStream(this.binary, args, this.spawnFn, {
+      parseLine: parseStreamJsonLine,
+      parseCtx: { runId: request.runId, now: this.now, onUnknown: this.onUnknown },
+    });
   }
 }
